# Conflicts:
#	Assets/GameScene.unity
#	Assets/StreamingAssets/ArbaitData.json
#	Assets/StreamingAssets/BossPanelInfo.json
#	Assets/StreamingAssets/CreateWeapon.json
#	Assets/StreamingAssets/Equiment.json
#	Assets/StreamingAssets/PlayerData.json
#	Assets/StreamingAssets/Quest.json
#	Library/CurrentMaximizeLayout.dwlt
#	Library/ScriptAssemblies/Assembly-CSharp.dll
#	Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	Library/assetDatabase3
#	Library/metadata/00/00000000000000004000000000000000
#	Library/metadata/18/18b4dbf8a0b54ea4adf46fb7f4d71dd0
#	Library/metadata/24/2403d1f95ea54028853403e595bc9274
#	Library/metadata/51/51288a4fc4384861a5b1f9dd49b3da26
#	Library/metadata/bb/bb6b4101ad2c42ec99f064a70c354e3c

+++ Assets/Scripts/InGame/UI/Enhance/AccuracyPanelUI.cs
@@ -32,7 +32,7 @@ public class AccuracyPanelUI : EnhanceUI {
 
 			fEnhanceValue = cPlayer.GetBasicAccuracyRate() + (fBasic * 0.01f);
 
-			cPlayer.SetBasicAccuracyRate(fEnhanceValue);
+			cPlayer.SetBasicAccuracyRate((float)fEnhanceValue);
 
 			EnhanceText.text =string.Format("{0} {1}", strEnhanceName , nLevel);
