+++ Assets/Scripts/InGame/UI/Enhance/WaterPlusPanelUI.cs
@@ -32,7 +32,7 @@ public class WaterPlusPanelUI : EnhanceUI {
 
 			fEnhanceValue = cPlayer.GetBasicWaterPlus() + (fBasic * 0.01f);
 
-			cPlayer.SetBasicWaterPlus(fEnhanceValue);
+			cPlayer.SetBasicWaterPlus((float)fEnhanceValue);
 
 			NextPercentText.text = (cPlayer.GetBasicWaterPlus() + (fBasic * 0.01f)).ToString();
